\name{pytha-package}
\alias{pytha-package}
\alias{pytha}
\docType{package}
\title{
\packageTitle{pytha}
}
\description{
\packageDescription{pytha}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pytha}
\packageIndices{pytha}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pytha}

Maintainer: \packageMaintainer{pytha}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
